feat(admin): implement room and user management controllers
- Add RoomAdminController with CRUD operations for room management:
  - Create, read, update, and delete rooms
  - Admin-only access enforcement
- Add UserAdminController with user administration features:
  - List all users with pagination
  - Update user roles/permissions
  - Enable/disable user accounts
  - Admin-only access enforcement

+++ src/main/java/com/axconstantino/reservationsystem/rooms/database/model/Room.java
@@ -6,6 +6,8 @@ import jakarta.persistence.*;
 import jakarta.validation.constraints.Size;
 import lombok.*;
 
+import java.math.BigDecimal;
+
 @Entity
 @Table(name = "rooms")
 @AllArgsConstructor
@@ -22,7 +24,7 @@ public class Room {
     private String name;
 
     @Column(nullable = false)
-    private Double pricePerNight;
+    private BigDecimal pricePerNight;
 
     @Column(nullable = false)
     private Integer capacity;
